+++ src/elite/test/EliteSharpLib.Tests/HeadlessGameHarness.cs
@@ -42,7 +42,7 @@ internal sealed class HeadlessGameHarness : HeadlessGameHarnessBase<GameStateSum
         services.AddSingleton(sp => sp.GetRequiredService<IAbstraction>().Graphics);
         services.AddSingleton(sp => sp.GetRequiredService<IAbstraction>().Sound);
         services.AddSingleton(sp => sp.GetRequiredService<IAbstraction>().Keyboard);
-        services.AddSingleton(_ => AssetLocator.Create());
+        services.AddSingleton<IAssetLocator>(_ => AssetLocator.Create());
         services.AddEliteConfig(_configDirectory);
         services.AddEliteMain();