+++ src/useful/libs/Useful.App/LogMessages.cs
@@ -1,10 +1,8 @@
-// 'Stunt Car Racer - The Sharp Kind' - Andy Hawkins 2026.
-// 'Stunt Car Racer Remake' - sourceforge.net/projects/stuntcarremake.
-// Stunt Car Racer (C) Geoff Crammond / MicroStyle / MicroProse 1989.
+// 'Useful Libraries' - Andy Hawkins 2023-2026.
 
 using Microsoft.Extensions.Logging;
 
-namespace StuntCarRacerSharp;
+namespace Useful.App;
 
 internal static partial class LogMessages
 {